pyproject.toml
[project]
name = "dataanalysis"
version = "0.1.0"
description = "Aplikacja do analizy danych magazynowych - pojemnosciowa i wydajnosciowa"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "DataAnalysis Team"}
]
keywords = ["warehouse", "analytics", "capacity", "performance", "kardex"]

dependencies = [
    "polars>=1.0.0",
    "duckdb>=1.0.0",
    "streamlit>=1.35.0",
    "pydantic>=2.0.0",
    "pyarrow>=15.0.0",
    "openpyxl>=3.1.0",
    "pyyaml>=6.0.0",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
dataanalysis = "src.ui.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
